+++ RELEASE_NOTES.md
@@ -1,9 +1,11 @@
 ### 2.0-dev - Unreleased
+* (Breaking change) Update.run signature changed
 * (Breaking change) SimpleUpdate.read moved to Update module
 * (Breaking change) SimpleUpdate.get is now Update.getState
 * Add nameOf, instanceOf operators
 * Range validation
 * Additional Result operators, sequence function and Builder methods
+* UpdateResult
 
 ### 1.0 - 2018/12/16
 * Un/curry and flip