+++ RELEASE_NOTES.md
@@ -1,4 +1,4 @@
-### 2.0-dev - Unreleased
+### 2.0.0-dev - Unreleased
 * (Breaking change) Update.run signature changed
 * (Breaking change) SimpleUpdate.read moved to Update module
 * (Breaking change) SimpleUpdate.get is now Update.getState
@@ -7,6 +7,7 @@
 * Additional Result operators, sequence function and Builder methods
 * UpdateResult
 * Improved TryWith, TryFinally and Using Builder methods
+* Add SourceLink support
 
 ### 1.0 - 2018/12/16
 * Un/curry and flip